+++ src/ws_server/controllers/games/finishGame.ts
@@ -1,10 +1,10 @@
-import { gamesStorage } from '../..';
 import { BOT_ID } from '../../constants';
 import { GameId } from '../../types';
 import { db } from '../../models/DataBase';
 import { GameBot } from '../../models/GameBot';
-import { updateWinners } from '../players/updateWinners';
+import { updateWinners } from '../players';
 import { finishGameBroadcast } from './broadcasters';
+import { gamesStorage } from '.';
 
 export const finishGame = (gameId: GameId) => {
   const gameInfo = gamesStorage.get(gameId);
