+++ src/ws_server/controllers/games/singlePlay.ts
@@ -1,11 +1,11 @@
 import { WebSocket } from 'ws';
-import { gamesStorage, registeredClients } from '../..';
 import { BattleShipGame } from '../../models/BattleShipGame';
 import { BOT_ID, GAME_FIELD_SIZE } from '../../constants';
 import { GameBot } from '../../models/GameBot';
 import { createGameBroadcast } from './broadcasters';
-import { deleteRoomWithUser } from '../rooms';
-import { updateRoom } from '../rooms/updateRoom';
+import { deleteRoomWithUser, updateRoom } from '../rooms';
+import { registeredClients } from '../players';
+import { gamesStorage } from '.';
 
 export const singlePlay = (ws: WebSocket) => {
   const player = registeredClients.get(ws);
@@ -14,8 +14,7 @@ export const singlePlay = (ws: WebSocket) => {
   deleteRoomWithUser(index);
   updateRoom();
 
-  const idGame = gamesStorage.size;
-  const newGame = new BattleShipGame(idGame, GAME_FIELD_SIZE);
+  const newGame = new BattleShipGame(GAME_FIELD_SIZE);
   const bot = new GameBot(newGame, ws);
 
   const players = new Map();
@@ -24,6 +23,6 @@ export const singlePlay = (ws: WebSocket) => {
   newGame.addPlayer(index);
   newGame.addPlayer(BOT_ID);
 
-  gamesStorage.set(idGame, { game: newGame, players });
-  createGameBroadcast(ws, idGame, index);
+  gamesStorage.set(newGame.gameId, { game: newGame, players });
+  createGameBroadcast(ws, newGame.gameId, index);
 };
