+++ src/ws_server/controllers/players/index.ts
@@ -1,0 +1,7 @@
+import { regPlayer } from './regPlayer';
+import { updateWinners } from './updateWinners';
+import { handleDisconnectedPlayer } from './handleDisconnectedPlayer';
+import { ClientsMap } from '../../types';
+
+export const registeredClients: ClientsMap = new Map();
+export { regPlayer, updateWinners, handleDisconnectedPlayer };
