+++ src/ws_server/controllers/rooms/index.ts
@@ -1,3 +1,9 @@
+import { createRoom } from './createRoom';
+import { addUserToRoom } from './addUserToRoom';
+import { updateRoom } from './updateRoom';
 import { deleteRoomWithUser } from './deleteRoomWithUser';
+import { RoomsMap } from '../../types';
 
-export { deleteRoomWithUser };
+export const roomsStorage: RoomsMap = new Map();
+
+export { createRoom, addUserToRoom, updateRoom, deleteRoomWithUser };
